Tests/LibWeb/Text/expected/geometry/getBoundingClientRect-empty-inline-vertical-writing-mode.txt

empty-in-horizontal: 0x20
empty-in-vertical: 0x20
